Making examples not rely on relative import path.
